+++ R/BTm.setup.R
@@ -1,13 +1,14 @@
 #' @importFrom stats reformulate
-BTm.setup <- function(outcome = 1, player1, player2, formula = NULL,
-                            id = "..", separate.ability = NULL, refcat = NULL,
-                            data = NULL, weights = NULL, subset = NULL,
-                            offset = NULL, contrasts = NULL, ...){
+BTm.setup <- function(player1, player2 = NULL, outcome = 1, formula = NULL,
+                      id = "..", separate.ability = NULL, refcat = NULL,
+                      data = NULL, weights = NULL, subset = NULL,
+                      offset = NULL, contrasts = NULL, ...){
     if (!is.data.frame(data)){
         keep <- names(data) %in% c(deparse(substitute(player1)),
                                    deparse(substitute(player2)))
         if (!length(keep)) keep <- FALSE
         ## save row names for checking against index variables (in Diff)
+        data <- lapply(data, as.data.frame)
         nm <- lapply(data, rownames)
         data <- c(data[keep], unlist(unname(data[!keep]), recursive = FALSE))
         if (any(dup <- duplicated(names(data))))
@@ -24,8 +25,11 @@ BTm.setup <- function(outcome = 1, player1, player2, formula = NULL,
         else eval(expr)
     }
     player1 <- withIfNecessary(substitute(player1), data)
+    if (ncol(player1) == 1) colnames(player1) <- id
+    if (ncol(player1) == 0) player1 <- data[id]
     player2 <- withIfNecessary(substitute(player2), data)
-    if (ncol(player1) == 1) colnames(player1) <- colnames(player2) <- id
+    if (ncol(player2) == 1) colnames(player2) <- id
+    if (ncol(player2) == 0) player2 <- NULL
     Y <- withIfNecessary(substitute(outcome), c(player1, player2, data),
                          as.data.frame = FALSE)
     weights <- withIfNecessary(substitute(weights), data, FALSE)